Reticular Action Model - The \( \mathbf{S} \) Matrix and The Model-Implied Variance Covariance Matrix \( \boldsymbol{\Sigma} \left( \boldsymbol{\theta} \right) %(\#eq:sem-Sigmatheta) \)

Derives the \( \mathbf{S} \) matrix and the model-implied variance-covariance matrix \( \boldsymbol{\Sigma} \left( \boldsymbol{\theta} \right) %(\#eq:sem-Sigmatheta) \) using the Reticular Action Model (RAM) notation.

ram_S(A, sigma2, F, I, SigmaMatrix = TRUE, filter = TRUE, both = FALSE)

Arguments

A
\(\mathbf{A}\) matrix. Asymmetric paths (single-headed arrows), such as regression coefficients and factor loadings.

sigma2
Numeric vector. Vector of variances \(\sigma^2\). The first element should be the variance of an exogenous variable.

F
\(\mathbf{F}\) matrix. Filter matrix used to select the observed variables.

I
\(\mathbf{I}\) matrix. Identity matrix.

SigmaMatrix
Logical. If TRUE, returns the model-implied variance-covariance matrix \( \boldsymbol{\Sigma} \left( \boldsymbol{\theta} \right) %(\#eq:sem-Sigmatheta) \) . If FALSE, returns the \( \mathbf{S} \) matrix.

filter
Logical. If TRUE, the F matrix is used. If FALSE, the F matrix is omitted and the argument F is ignored. See Value and Details below for more information.

both
Logical. If TRUE, returns both the \( \mathbf{S} \) marix and the model-implied variance-covariance matrix \( \boldsymbol{\Sigma} \left( \boldsymbol{\theta} \right) %(\#eq:sem-Sigmatheta) \) in a list.

Value

If both = TRUE, returns both the \( \mathbf{S} \) matrix and the model-implied variance-covariance matrix \( \boldsymbol{\Sigma} \left( \boldsymbol{\theta} \right) %(\#eq:sem-Sigmatheta) \) derived from the \(\mathbf{A}\) matrix and \(\sigma^2\) vector. in a list. Otherwise, If both = FALSE, the model-implied variance-covariance matrix \( \boldsymbol{\Sigma} \left( \boldsymbol{\theta} \right) %(\#eq:sem-Sigmatheta) \) is returned if SigmaMatrix = TRUE. If both = FALSE, \( \mathbf{S} \) matrix is returned if SigmaMatrix = FALSE.

NOTE: SigmaMatrix - option to filter out latent variables If filter = TRUE, variance-covariance matrix of observed variables. If filter = FALSE, combined variance-covariance matrix of observed and latent variables. S - always the full S matrix with the same dimensions as the input A matrix

Details

The \( \mathbf{S} \) matrix and the model-implied variance-covariance matrix \( \boldsymbol{\Sigma} \left( \boldsymbol{\theta} \right) %(\#eq:sem-Sigmatheta) \) are derived using the \(\mathbf{A}\) matrix and sigma squared \(\left( \sigma^2 \right)\) vector (variances). Note that the first element in the A ans S matrices should be an exogenous variable.

References

McArdle, J. J. (2013). The development of the RAM rules for latent variable structural equation modeling. In A. Maydeu-Olivares & J. J. McArdle (Eds.), Contemporary Psychometrics: A festschrift for Roderick P. McDonald (pp. 225--273). Lawrence Erlbaum Associates.

McArdle, J. J., & McDonald, R. P. (1984). Some algebraic properties of the Reticular Action Model for moment structures. British Journal of Mathematical and Statistical Psychology, 37 (2), 234--251.

See also

Examples

A <- matrix(
  data = c(
    0, 0.26^(1 / 2), 0,
    0, 0, 0.26^(1 / 2),
    0, 0, 0
  ),
  ncol = 3
)
sigma2 <- c(15^2, 15^2, 15^2)
F <- I <- diag(3)
# Returns the model-implied variance-covariance matrix
ram_S(A = A, sigma2 = sigma2, F = F, I = I, SigmaMatrix = TRUE)

#>          [,1]     [,2]     [,3]
#> [1,] 225.0000 114.7279  58.5000
#> [2,] 114.7279 225.0000 114.7279
#> [3,]  58.5000 114.7279 225.0000

# Returns the model-implied S matrix
ram_S(A = A, sigma2 = sigma2, F = F, I = I, SigmaMatrix = FALSE)

#>      [,1]  [,2]  [,3]
#> [1,]  225   0.0   0.0
#> [2,]    0 166.5   0.0
#> [3,]    0   0.0 166.5

# One-factor CFA model--------------------------------------------------------
A <- matrix(data = 0, ncol = 6, nrow = 6)
for (i in 2:6) {
  A[i, 1] <- 0.5
}
sigma2 <- c(1, 1, 1, 1, 1, 1)
F <- I <- diag(nrow(A))
F <- diag(nrow(A) - 1)
F <- cbind(0, F)
ram_S(
  A = A,
  sigma2 = sigma2,
  F = F,
  I = I
)

#>      [,1] [,2] [,3] [,4] [,5]
#> [1,] 1.00 0.25 0.25 0.25 0.25
#> [2,] 0.25 1.00 0.25 0.25 0.25
#> [3,] 0.25 0.25 1.00 0.25 0.25
#> [4,] 0.25 0.25 0.25 1.00 0.25
#> [5,] 0.25 0.25 0.25 0.25 1.00

ram_S(
  A = A,
  sigma2 = sigma2,
  F = F,
  I = I,
  SigmaMatrix = FALSE
)

#>      [,1] [,2] [,3] [,4] [,5] [,6]
#> [1,]    1 0.00 0.00 0.00 0.00 0.00
#> [2,]    0 0.75 0.00 0.00 0.00 0.00
#> [3,]    0 0.00 0.75 0.00 0.00 0.00
#> [4,]    0 0.00 0.00 0.75 0.00 0.00
#> [5,]    0 0.00 0.00 0.00 0.75 0.00
#> [6,]    0 0.00 0.00 0.00 0.00 0.75