Reticular Action Model - Model-Implied Variance Covariance Matrix \( \boldsymbol{\Sigma} \left( \boldsymbol{\theta} \right) %(\#eq:sem-Sigmatheta) \)

Derives the model-implied variance-covariance matrix \( \boldsymbol{\Sigma} \left( \boldsymbol{\theta} \right) %(\#eq:sem-Sigmatheta) \) using the Reticular Action Model (RAM) notation.

ram_Sigmatheta(A, S, F, I, filter = TRUE)

Arguments

A
\(\mathbf{A}\) matrix. Asymmetric paths (single-headed arrows), such as regression coefficients and factor loadings.

S
\(\mathbf{S}\) matrix. Symmetric paths (double-headed arrows), such as variances and covariances.

F
\(\mathbf{F}\) matrix. Filter matrix used to select the observed variables.

I
\(\mathbf{I}\) matrix. Identity matrix.

filter
Logical. If TRUE, the F matrix is used. If FALSE, the F matrix is omitted and the argument F is ignored. See Value and Details below for more information.

Value

Returns the model-implied variance-covariance matrix \(\left( \boldsymbol{\Sigma} \left( \boldsymbol{\theta} \right) \right)\) derived from the \(\mathbf{A}\), \(\mathbf{S}\), \(\mathbf{F}\), and \(\mathbf{I}\) matrices. If filter = TRUE, the F matrix is used as usual. $$ \boldsymbol{\Sigma} \left( \boldsymbol{\theta} \right) = \mathbf{F} \left( \mathbf{I} - \mathbf{A} \right)^{-1} \mathbf{S} \left[ \left( \mathbf{I} - \mathbf{A} \right)^{-1} \right]^{\prime} \mathbf{F}^{\prime} %(\#eq:sem-ram-Sigmatheta) $$ If filter = FALSE, the F matrix is omitted. $$ \boldsymbol{\Sigma} \left( \boldsymbol{\theta} \right) = \left( \mathbf{I} - \mathbf{A} \right)^{-1} \mathbf{S} \left[ \left( \mathbf{I} - \mathbf{A} \right)^{-1} \right]^{\prime} . %(\#eq:sem-ram-Sigmatheta-filter-false) $$

Details

The model-implied variance-covariance matrix \( \boldsymbol{\Sigma} \left( \boldsymbol{\theta} \right) %(\#eq:sem-Sigmatheta) \) as a function of Reticular Action Model (RAM) matrices is given by $$ \boldsymbol{\Sigma} \left( \boldsymbol{\theta} \right) = \mathbf{F} \left( \mathbf{I} - \mathbf{A} \right)^{-1} \mathbf{S} \left[ \left( \mathbf{I} - \mathbf{A} \right)^{-1} \right]^{\prime} \mathbf{F}^{\prime} %(\#eq:sem-ram-Sigmatheta) $$ where \(\mathbf{A}\) represents asymmetric paths (single-headed arrows), such as regression coefficients and factor loadings. \(\mathbf{S}\) represents symmetric paths (double-headed arrows), such as variances and covariances. \(\mathbf{F}\) represents the filter matrix used to select the observed variables. \(\mathbf{I}\) represents the identity matrix.

When all the variables in the \(\mathbf{A}\) and \(\mathbf{S}\) matrices are observed, $$ \boldsymbol{\Sigma} \left( \boldsymbol{\theta} \right) = \mathbf{F} \left( \mathbf{I} - \mathbf{A} \right)^{-1} \mathbf{S} \left[ \left( \mathbf{I} - \mathbf{A} \right)^{-1} \right]^{\prime} \mathbf{F}^{\prime} %(\#eq:sem-ram-Sigmatheta) $$ is equl to $$ \boldsymbol{\Sigma} \left( \boldsymbol{\theta} \right) = \left( \mathbf{I} - \mathbf{A} \right)^{-1} \mathbf{S} \left[ \left( \mathbf{I} - \mathbf{A} \right)^{-1} \right]^{\prime} , \\ \quad \text{when} \quad \mathbf{F} = \mathbf{I} . %(\#eq:sem-ram-Sigmatheta-observed) $$

References

McArdle, J. J. (2013). The development of the RAM rules for latent variable structural equation modeling. In A. Maydeu-Olivares & J. J. McArdle (Eds.), Contemporary Psychometrics: A festschrift for Roderick P. McDonald (pp. 225--273). Lawrence Erlbaum Associates.

McArdle, J. J., & McDonald, R. P. (1984). Some algebraic properties of the Reticular Action Model for moment structures. British Journal of Mathematical and Statistical Psychology, 37 (2), 234--251.

See also

Examples

# Simple mediation model with observed variables------------------------------
# parameters
alpha <- 0.5
beta <- 0.4
tauprime <- 0.25
sigma2X <- 1.00
sigma2epsilonM <- 0.75
sigma2epsilonY <- 0.6775
# A and S matrices only contain observed variables X, M, and Y
A <- matrix(data = 0, nrow = 3, ncol = 3)
A[2, 1] <- alpha
A[3, 1] <- beta
A[3, 2] <- tauprime
S <- diag(c(sigma2X, sigma2epsilonM, sigma2epsilonY))
F <- I <- diag(3)
ram_Sigmatheta(A = A, S = S, F = F, I = I)

#>       [,1] [,2]  [,3]
#> [1,] 1.000 0.50 0.525
#> [2,] 0.500 1.00 0.450
#> [3,] 0.525 0.45 1.000

# Equivalent
ram_Sigmatheta(A = A, S = S, I = I, filter = FALSE)

#>       [,1] [,2]  [,3]
#> [1,] 1.000 0.50 0.525
#> [2,] 0.500 1.00 0.450
#> [3,] 0.525 0.45 1.000

# A and S matrices also contain latent variables sigma2epsilonM and sigma2epsilonY
A <- matrix(data = 0, nrow = 5, ncol = 5)
A[2, 1] <- alpha
A[2, 4] <- 1
A[3, 1] <- beta
A[3, 2] <- tauprime
A[3, 5] <- 1
S <- diag(c(sigma2X, 0, 0, sigma2epsilonM, sigma2epsilonY))
I <- diag(nrow(A))
F <- matrix(data = 0, nrow = 3, ncol = 5)
F[1, 1] <- 1
F[2, 2] <- 1
F[3, 3] <- 1
# In this case since F is not equal to I, we CANNOT use filter = FALSE
ram_Sigmatheta(A = A, S = S, F = F, I = I)

#>       [,1] [,2]  [,3]
#> [1,] 1.000 0.50 0.525
#> [2,] 0.500 1.00 0.450
#> [3,] 0.525 0.45 1.000

# One-factor CFA model--------------------------------------------------------
A <- matrix(data = 0, ncol = 6, nrow = 6)
for (i in 2:6) {
  A[i, 1] <- 0.5
}
S <- diag(c(1, 0.75, 0.75, 0.75, 0.75, 0.75))
F <- I <- diag(nrow(A))
F <- diag(nrow(A) - 1)
F <- cbind(0, F)
ram_Sigmatheta(
  A = A,
  S = S,
  F = F,
  I = I
)

#>      [,1] [,2] [,3] [,4] [,5]
#> [1,] 1.00 0.25 0.25 0.25 0.25
#> [2,] 0.25 1.00 0.25 0.25 0.25
#> [3,] 0.25 0.25 1.00 0.25 0.25
#> [4,] 0.25 0.25 0.25 1.00 0.25
#> [5,] 0.25 0.25 0.25 0.25 1.00