Reticular Action Model - Model-Implied Mean Vector \( \boldsymbol{\mu} \left( \boldsymbol{\theta} \right)\)

Derives the model-implied mean vector \( \boldsymbol{\mu} \left( \boldsymbol{\theta} \right)\) using the Reticular Action Model (RAM) notation.

ram_mutheta(M, A, filter)

Arguments

M
p x 1 numeric vector \(\mathbf{M}_{p \times 1}\). Mean structure. Vector of means and intercepts.

A
p x p numeric matrix \(\mathbf{A}_{p \times p}\). Asymmetric paths (single-headed arrows), such as regression coefficients and factor loadings.

filter
k x p numeric matrix \(\mathbf{F}_{k \times p}\). Filter matrix used to select variables.

Value

Returns the model-implied mean vector \( \boldsymbol{\mu} \left( \boldsymbol{\theta} \right)\) derived from the \(\mathbf{A}\), \(\mathbf{F}\), \(\mathbf{I}\), matrices and \(\mathbf{M}\) vector.

Details

The model-implied mean vector \( \boldsymbol{\mu} \left( \boldsymbol{\theta} \right)\) as a function of Reticular Action Model (RAM) matrices is given by $$ \boldsymbol{\mu} \left( \boldsymbol{\theta} \right) = \mathbf{F} \left( \mathbf{I} - \mathbf{A} \right)^{-1} \mathbf{M}$$ where \(\mathbf{M}\) is the \(p \times 1\) mean structure, that is, a vector of means and intercepts.

References

McArdle, J. J. (2013). The development of the RAM rules for latent variable structural equation modeling. In A. Maydeu-Olivares & J. J. McArdle (Eds.), Contemporary Psychometrics: A festschrift for Roderick P. McDonald (pp. 225--273). Lawrence Erlbaum Associates.

McArdle, J. J., & McDonald, R. P. (1984). Some algebraic properties of the Reticular Action Model for moment structures. British Journal of Mathematical and Statistical Psychology, 37 (2), 234--251.

See also

Examples

#-----------------------------------------------------------------------------
# Simple mediation model with observed variables------------------------------
# The variables in the model are X, M, and Y.
# All three are observed variables.
#-----------------------------------------------------------------------------
alpha <- 0.3385926
beta <- 0.4510391
tauprime <- 0.2076475
muX <- 70.18000
deltaM <- -20.70243
deltaY <- -12.71288
A <- matrix(data = 0, nrow = 3, ncol = 3)
A[2, 1] <- alpha
A[3, 1] <- beta
A[3, 2] <- tauprime
filter <- diag(3)
M <- c(X = muX, M = deltaM, Y = deltaY)
colnames(A) <- c("X", "M", "Y")
rownames(A) <- c("X", "M", "Y")
colnames(filter) <- c("X", "M", "Y")
rownames(filter) <- c("X", "M", "Y")
ram_mutheta(M = M, A = A, filter = filter)

#>        [,1]
#> X 70.180000
#> M  3.059999
#> Y 19.576445

#-----------------------------------------------------------------------------
# Simple mediation model with observed variables------------------------------
# and latent variable error terms---------------------------------------------
# The same mediation model can be parameterized by explicitly
# including the error terms epsilonM and epsilonY in the matrices.
# The variables in the model are X, M, Y, epsilonM, and epsilonY.
# X, M, and Y are observed variables.
# epsilonM, and epsilonY are latent variables.
#-----------------------------------------------------------------------------
A <- matrix(data = 0, nrow = 5, ncol = 5)
A[2, 1] <- alpha
A[2, 4] <- 1
A[3, 1] <- beta
A[3, 2] <- tauprime
A[3, 5] <- 1
filter <- matrix(data = 0, nrow = 3, ncol = 5)
filter[1, 1] <- 1
filter[2, 2] <- 1
filter[3, 3] <- 1
M <- c(X = muX, M = deltaM, Y = deltaY, epsilonM = 0, epsilonY = 0)
colnames(A) <- c("X", "M", "Y", "epsilonM", "epsilonY")
rownames(A) <- c("X", "M", "Y", "epsilonM", "epsilonY")
colnames(filter) <- c("X", "M", "Y", "epsilonM", "epsilonY")
rownames(filter) <- c("X", "M", "Y")
ram_mutheta(M = M, A = A, filter = filter)

#>        [,1]
#> X 70.180000
#> M  3.059999
#> Y 19.576445

# One-factor CFA model--------------------------------------------------------
# The variables in the model are eta, y1, y2, y3, y4, y5.
# eta is a latent variable. y1 to y5 are observed variables.
#-----------------------------------------------------------------------------
A <- matrix(data = 0, ncol = 6, nrow = 6)
loadings <- 0.5 # tau equivalence
for (i in 2:6) {
  A[i, 1] <- 0.5
}
filter <- diag(nrow(A) - 1)
filter <- cbind(0, filter)
M <- c(eta = 0, y1 = 0, y2 = 0, y3 = 0, y4 = 0, y5 = 0)
colnames(A) <- c("eta", "y1", "y2", "y3", "y4", "y5")
rownames(A) <- c("eta", "y1", "y2", "y3", "y4", "y5")
colnames(filter) <- c("eta", "y1", "y2", "y3", "y4", "y5")
rownames(filter) <- c("y1", "y2", "y3", "y4", "y5")
ram_mutheta(M = M, A = A, filter = filter)

#>    [,1]
#> y1    0
#> y2    0
#> y3    0
#> y4    0
#> y5    0